changes pdf

+++ application/views/employee/pages/view/pendinglist.php
@@ -22,14 +22,14 @@
               </tr>
             </thead>
           <tbody>
-              <?php foreach ($pendingdetails as $key) { ?>
+              <?php if(!empty($pendingdetails)) { foreach($pendingdetails as $key) {?>
                 <tr>
                   <td><a href="pendingupdate?lead_code=<?php echo $key->lead_code;?>"><i class="fa fa-pencil-square-o"></i>Edit</a><a href="pendingdelete?lead_code=<?php echo $key->lead_code;?>"><i class="fa fa fa-trash" style="margin:3px;"></i>Delete</a></td>
                   <td><?php echo $key->id;?></td>
                   <td><?php echo client_name($key->lead_code);?></td>
                   <td><?php echo __date_format($key->created_at,'ddmmyyyy');?></td>
                 </tr> 
-              <?php } ?>
+              <?php } } ?>
           </tbody>
           </table>
         </div>
